Add conditionals for package install

+++ install.sh
@@ -5,19 +5,24 @@ ZABBIX_PACKAGE_URL="http://repo.zabbix.com/zabbix/${ZABBIX_VERSION}/ubuntu/pool/
 ZABBIX_PACKAGE_NAME="zabbix-release_${ZABBIX_RELEASE}+`lsb_release -c -s`_all.deb"
 
 if [ "$EUID" -ne 0 ]; then 
-	echo "Please run as root"
-	exit 1
+  echo "Please run as root"
+  exit 1
 fi
 
 if ! dpkg -l |grep -q zabbix-sender; then
-	wget ${ZABBIX_PACKAGE_URL}/${ZABBIX_PACKAGE_NAME} -O /tmp/${ZABBIX_PACKAGE_NAME}
-	dpkg -i /tmp/${ZABBIX_PACKAGE_NAME}
-	apt-get update
-	apt-get install -y zabbix-sender
+  wget ${ZABBIX_PACKAGE_URL}/${ZABBIX_PACKAGE_NAME} -O /tmp/${ZABBIX_PACKAGE_NAME}
+  dpkg -i /tmp/${ZABBIX_PACKAGE_NAME}
+  apt-get update
+  apt-get install -y zabbix-sender
 fi
 
-apt-get install -y python-pip python-dev 
-pip install mysql-connector-python --allow-external mysql-connector-python
+if ! dpkg -l |egrep -q 'python-pip|python-dev'; then
+  apt-get install -y python-pip python-dev
+fi
+
+if ! dpkg -l |grep -q 'mysql-connector-python'; then
+  pip install mysql-connector-python --allow-external mysql-connector-python
+fi
 
 ./build.sh
 ./setup.py install --force
